Spliced the analyzer.js file into separate files for each prototype to reduce complexity
