+++ html/score.html
@@ -6,6 +6,9 @@
         <script type="text/javascript" src="../js/config/config.js"></script>
         <script type="text/javascript" src="../js/config/scoring.js"></script>
         <script type="text/javascript" src="../js/analyzer.js"></script>
+        <script type="text/javascript" src="../js/analyzescorer.js"></script>
+        <script type="text/javascript" src="../js/preprocessor.js"></script>
+        <script type="text/javascript" src="../js/stringhelper.js"></script>
         <script type="text/javascript" src="../js/outputter.js"></script>
 
         <link rel="stylesheet" href="../css/style.css" type="text/css">
