+++ js/analyzer.js
@@ -491,446 +491,3 @@ Analyzer.prototype.urlStopwords = function() {
 Analyzer.prototype.score = function() {
     this.analyzeScorer.score(this.__output);
 };
-
-/**
- * PreProcessor object definition. Creates __store object and calls init.
- * @params textString
- */
-PreProcessor = function ( text ){
-    //create __store object to store data
-    this.__store = {};
-    this.__store.originalText = text;
-    this.stringHelper = stringHelper();
-    this.init();
-};
-
-/**
- * init function calling all necessary PreProcessorfunctions
- */
-PreProcessor.prototype.init = function() {
-    //call function to clean text
-    this.textFormat();
-    //call function to count words
-    this.countStore();
-};
-
-/**
- * formats the original text from __store and save as cleantext, cleantextSomeTags en cleanTextNoTags
- */
-PreProcessor.prototype.textFormat = function() {
-    this.__store.cleanText = this.cleanText( this.__store.originalText );
-    this.__store.cleanTextSomeTags = this.stripSomeTags( this.__store.cleanText );
-    this.__store.cleanTextNoTags = this.stripAllTags( this.__store.cleanTextSomeTags );
-};
-
-/**
- * saves wordcount (all words) and wordcountNoTags (all words except those in tags) in the __store object
- */
-PreProcessor.prototype.countStore = function() {
-    /*wordcounters*/
-    this.__store.wordcount = this.__store.cleanText === "." ? 0 : this.__store.cleanText.split( " " ).length;
-    this.__store.wordcountNoTags = this.__store.cleanTextNoTags.split( " " ).length;
-    /*sentencecounters*/
-    this.__store.sentenceCount = this.sentenceCount( this.__store.cleanText );
-    this.__store.sentenceCountNoTags = this.sentenceCount( this.__store.cleanTextNoTags );
-    /*syllablecounters*/
-    this.__store.syllablecount = this.syllableCount( this.__store.cleanTextNoTags );
-};
-
-/**
- * counts the number of sentences in a textstring by splitting on a period. Removes sentences that are empty or have only a space.
- * @param textString
- */
-PreProcessor.prototype.sentenceCount = function( textString ){
-    var sentences = textString.split( "." );
-    sentenceCount = 0;
-    for ( var i = 0; i < sentences.length; i++ ){
-        if( sentences[i] !== "" && sentences[i] !== " " ){
-            sentenceCount++;
-        }
-    }
-    return sentenceCount;
-};
-
-/**
- * counts the number of syllables in a textstring, calls exclusionwordsfunction, basic syllable counter and advanced syllable counter.
- * @param textString
- * @returns syllable count
- */
-PreProcessor.prototype.syllableCount = function( textString ) {
-    this.syllableCount = 0;
-    textString = textString.replace( /[.]/g, " " );
-    textString = this.removeWords( textString );
-    var words = textString.split( " " );
-    var subtractSyllablesRegexp = this.stringHelper.stringToRegex( preprocessorConfig.syllables.subtractSyllables, true );
-    var addSyllablesRegexp = this.stringHelper.stringToRegex( preprocessorConfig.syllables.addSyllables, true );
-    for ( var i = 0; i < words.length; i++ ){
-        this.basicSyllableCount( words[i].split( /[^aeiouy]/g ) );
-        this.advancedSyllableCount( words[i], subtractSyllablesRegexp, "subtract" );
-        this.advancedSyllableCount( words[i], addSyllablesRegexp, "add" );
-    }
-    return this.syllableCount;
-};
-
-/**
- * counts the syllables by splitting on consonants
- * @param splitWordArray
- */
-
-PreProcessor.prototype.basicSyllableCount = function( splitWordArray ) {
-    for ( var j = 0; j < splitWordArray.length; j++ ){
-        if( splitWordArray[j].length > 0 ){
-            this.syllableCount++;
-        }
-    }
-};
-
-/**
- * counts the syllables by validating against regexxes, and adding and subtracting the number of matches.
- * @param inputString
- * @param regex
- * @param operator
- */
-PreProcessor.prototype.advancedSyllableCount = function( inputString, regex, operator ) {
-    var match = inputString.match( regex );
-    if( match !== null ){
-        if( operator === "subtract" ){
-            this.syllableCount -= match.length;
-        }else if( operator === "add" ){
-            this.syllableCount += match.length;
-        }
-    }
-};
-
-/**
- * removes words from textstring and count syllables. Used for words that fail against regexes.
- * @param textString
- * @returns textString with exclusionwords removed
- */
-PreProcessor.prototype.removeWords = function( textString ) {
-    for ( var i = 0; i < preprocessorConfig.syllables.exclusionWords.length; i++ ){
-        var exclusionRegex = new RegExp( preprocessorConfig.syllables.exclusionWords[i].word, "g" );
-        var matches = textString.match( exclusionRegex );
-        if( matches !== null ){
-            this.syllableCount += preprocessorConfig.syllables.exclusionWords[i].syllables;
-            textString = textString.replace( exclusionRegex, "" );
-        }
-    }
-    return textString;
-};
-
-/**
- * cleans text by removing special characters, numberonly words and replacing all terminators by periods
- * @param textString
- * @returns textString
- */
-PreProcessor.prototype.cleanText = function( textString ) {
-    textString = textString.toLocaleLowerCase();
-    //replace comma', hyphens etc with spaces
-    textString = textString.replace( /[\-\;\:\,\(\)\"\'\|\“\”]/g, " " );
-    //remove apostrophe
-    textString = textString.replace( /[\’]/g, "" );
-    //unify all terminators
-    textString = textString.replace( /[.?!]/g, "." );
-    //add period in case it is missing
-    textString += ".";
-    //replace newlines with spaces
-    textString = textString.replace( /[ ]*(\n|\r\n|\r)[ ]*/g, " " );
-    //remove duplicate terminators
-    textString = textString.replace( /([\.])[\. ]+/g, "$1" );
-    //pad sentence terminators
-    textString = textString.replace( /[ ]*([\.])+/g, "$1 " );
-    //Remove "words" comprised only of numbers
-    textString = textString.replace( /[0-9]+[ ]/g, "" );
-    return this.stringHelper.stripSpaces( textString );
-};
-
-/**
- * removes all HTMLtags from input string, except h1-6, li, p and dd
- * @param textString
- * @returns textString
- */
-PreProcessor.prototype.stripSomeTags = function( textString ) {
-    //remove tags, except li, p, h1-6, dd
-    textString = textString.replace( /<(?!li|\/li|p|\/p|h1|\/h1|h2|\/h2|h3|\/h3|h4|\/h4|h5|\/h5|h6|\/h6|dd).*?\>/g, " " );
-    textString = this.stringHelper.stripSpaces( textString );
-    return textString;
-};
-
-/**
- * remove all HTMLtags from input string.
- * @param textString
- * @returns textString
- */
-PreProcessor.prototype.stripAllTags = function( textString ) {
-    //remove all tags
-    textString = textString.replace( /(<([^>]+)>)/ig," " );
-    textString = this.stringHelper.stripSpaces( textString );
-    return textString;
-};
-/**
- * inits the analyzerscorer used for scoring of the output from the textanalyzer
- * @constructor
- */
-AnalyzeScorer = function( refObj ) {
-    this.__score = [];
-    this.refObj = refObj;
-    this.init();
-};
-
-/**
- * loads the analyzerScoring from the config file.
- */
-AnalyzeScorer.prototype.init = function() {
-    this.scoring = analyzerScoring;
-};
-
-/**
- * Starts the scoring by taking the resultObject from the analyzer. Then runs the scorequeue.
- * @param resultObj
- */
-AnalyzeScorer.prototype.score = function( resultObj ) {
-    this.resultObj = resultObj;
-    this.runQueue();
-};
-
-/**
- * runs the queue and saves the result in the __score-object.
- */
-AnalyzeScorer.prototype.runQueue = function() {
-    for ( var i = 0; i < this.resultObj.length; i++ ){
-        var subScore = this.genericScore( this.resultObj[i] );
-        if( typeof subScore !== "undefined" ) {
-            this.__score = this.__score.concat( subScore );
-        }
-    }
-    this.__totalScore = this.totalScore();
-};
-
-/**
- * scoring function that returns results based on the resultobj from the analyzer matched with
- * the scorearrays in the scoring config.
- * @param obj
- * @returns {{name: (analyzerScoring.scoreName), score: number, text: string}}
- */
-AnalyzeScorer.prototype.genericScore = function( obj ){
-    if( typeof obj !== "undefined" ) {
-        var scoreObj = this.scoreLookup( obj.test );
-        var score = { name: scoreObj.scoreName, score: 0, text: "" };
-        for ( var i = 0; i < scoreObj.scoreArray.length; i++ ) {
-            this.setMatcher( obj, scoreObj, i );
-            switch ( true ) {
-                case ( typeof scoreObj.scoreArray[i].type === "string" && this.result[scoreObj.scoreArray[i].type] ):
-                    return this.returnScore( score, scoreObj, i );
-                case ( typeof scoreObj.scoreArray[i].min === "undefined" && this.matcher <= scoreObj.scoreArray[i].max ):
-                    return this.returnScore( score, scoreObj, i );
-                case ( typeof scoreObj.scoreArray[i].max === "undefined" && this.matcher >= scoreObj.scoreArray[i].min ):
-                    return this.returnScore( score, scoreObj, i );
-                case ( this.matcher >= scoreObj.scoreArray[i].min && this.matcher <= scoreObj.scoreArray[i].max ):
-                    return this.returnScore( score, scoreObj, i );
-                default:
-                    break;
-            }
-        }
-        return score;
-    }
-};
-
-/**
- * sets matcher and resultvariables so the scorefunction can use this.
- * @param obj
- * @param scoreObj
- * @param i
- */
-AnalyzeScorer.prototype.setMatcher = function( obj, scoreObj, i ) {
-    this.matcher = parseFloat( obj.result );
-    this.result = obj.result;
-    if( typeof scoreObj.scoreArray[i].matcher !== "undefined" ){
-        this.matcher = parseFloat( this.result[scoreObj.scoreArray[i].matcher] );
-    }
-};
-
-/**
- * finds the scoringobject by scorename for the current result.
- * @param name
- * @returns scoringObject
- */
-AnalyzeScorer.prototype.scoreLookup = function( name ) {
-    for ( var ii = 0; ii < this.scoring.length; ii++ ){
-        if ( name === this.scoring[ii].scoreName ){
-            return this.scoring[ii];
-        }
-    }
-};
-
-/**
- * fills the score with score and text from the scoreArray and runs the textformatter.
- * @param score
- * @param scoreObj
- * @param i
- * @returns scoreObject
- */
-AnalyzeScorer.prototype.returnScore = function( score, scoreObj, i ) {
-    score.score = scoreObj.scoreArray[i].score;
-    score.text = this.scoreTextFormat( scoreObj.scoreArray[i], scoreObj.replaceArray );
-    return score;
-};
-
-/**
- * Formats the resulttexts with variables. Uses a value, source, sourceObj or scoreObj for the replacement source
- * replaces the position from the replaceArray with the replacement source.
- * @param scoreObj
- * @param replaceArray
- * @returns formatted resultText
- */
-AnalyzeScorer.prototype.scoreTextFormat = function( scoreObj, replaceArray ) {
-    var resultText = scoreObj.text;
-    if( typeof replaceArray !== "undefined" ) {
-        for ( var i = 0; i < replaceArray.length; i++ ) {
-            switch( true ) {
-                case ( typeof replaceArray[i].value !== "undefined" ):
-                    resultText = resultText.replace( replaceArray[i].position, replaceArray[i].value );
-                    break;
-                case ( typeof replaceArray[i].source !== "undefined" ):
-                    resultText = resultText.replace( replaceArray[i].position, this[replaceArray[i].source] );
-                    break;
-                case ( typeof replaceArray[i].sourceObj !== "undefined" ):
-                    var replaceWord = this.parseReplaceWord( replaceArray[i].sourceObj );
-                    resultText = resultText.replace( replaceArray[i].position, replaceWord );
-                    break;
-                case ( typeof replaceArray[i].scoreObj !== "undefined" ):
-                    resultText = resultText.replace( replaceArray[i].position, scoreObj[replaceArray[i].scoreObj] );
-                    break;
-                default:
-                    break;
-            }
-        }
-    }
-    return resultText;
-};
-
-/**
- * converts the string to the correct object and returns the string to be used in the text.
- * @param replaceWord
- * @returns {AnalyzeScorer}
- */
-AnalyzeScorer.prototype.parseReplaceWord = function( replaceWord ) {
-    var parts = replaceWord.split( "." );
-    var source = this;
-    for ( var i = 1; i < parts.length; i++ ){
-        source = source[ parts[i] ];
-    }
-    return source;
-};
-
-AnalyzeScorer.prototype.totalScore = function() {
-    var scoreAmount = this.__score.length;
-    var totalScore = 0;
-    for( var i = 0; i < this.__score.length; i++ ){
-        if( typeof this.__score[i] !== "undefined" ) {
-            totalScore += this.__score[i].score;
-        } else {
-            scoreAmount--;
-        }
-    }
-    var totalAmount = scoreAmount * analyzerScoreRating;
-    return Math.round( ( totalScore / totalAmount ) * 10 );
-};
-
-/**helper functions*/
-StringHelper = function() {};
-/**
- * removes strings from array and replaces them with keyword.
- * @param textString
- * @param stringsToRemove []
- * @param replacement (default == space)
- * @returns {textString}
- */
-StringHelper.prototype.replaceString = function( textString, stringsToRemove, replacement ){
-    if( typeof replacement === "undefined" ){
-        replacement = " ";
-    }
-    textString = textString.replace( this.stringToRegex(stringsToRemove), replacement );
-    return this.stripSpaces( textString );
-};
-
-/**
- * matches string with given array of strings to match.
- * @param textString
- * @param stringsToMatch
- * @returns {matches}
- */
-StringHelper.prototype.matchString = function( textString, stringsToMatch ){
-    return textString.match( this.stringToRegex( stringsToMatch ) );
-};
-
-/**
- * matches string with regex, returns the number of matches
- * @param textString
- * @param regex
- * @returns {number}
- */
-StringHelper.prototype.countMatches = function( textString, regex ){
-    var count = 0;
-    var matches = textString.match( regex );
-    if( matches !== null ){
-        count = matches.length;
-    }
-    return count;
-};
-
-/**
- * builds regex from array with strings
- * @param stringArray
- * @returns {RegExp}
- */
-StringHelper.prototype.stringToRegex = function( stringArray, disableWordBoundary ){
-    var regexString = "";
-    var wordBoundary = "\\b";
-    if( disableWordBoundary ){
-        wordBoundary = "";
-    }
-    for( var i = 0; i < stringArray.length; i++ ){
-        if(regexString.length > 0){ regexString += "|"; }
-        regexString += stringArray[i]+wordBoundary;
-    }
-    return new RegExp( regexString, "g" );
-};
-
-/**
- * Strip extra spaces, replace duplicates with single space. Remove space at front / end of string
- * @param textString
- * @returns textString
- */
-StringHelper.prototype.stripSpaces = function( textString ) {
-    //replace multiple spaces with single space
-    textString = textString.replace( / {2,}/g, " " );
-    //remove first/last character if space
-    textString = textString.replace( /^\s+|\s+$/g, "" );
-    return textString;
-};
-
-
-
-/**
- * Checks if the preprocessor is already initialized and if so if the textstring differs from the input.
- * @param inputString
- * @returns {yst_preProcessor}
- */
-preProcessor = function( inputString ) {
-    if ( typeof yst_preProcessor !== "object" || yst_preProcessor.inputText !== inputString ) {
-        yst_preProcessor = new PreProcessor( inputString );
-    }
-    return yst_preProcessor;
-};
-
-/**
- * Checks if the stringhelper is already initialized. Returns stringHelper.
- * @returns {StringHelper}
- */
-stringHelper = function() {
-    if ( typeof yst_stringHelper !== "object" ){
-        yst_stringHelper = new StringHelper();
-    }
-    return yst_stringHelper;
-};
