+++ spec/firstparagraphSpec.js
@@ -1,7 +1,9 @@
 require("../js/config/config.js");
 require("../js/config/scoring.js");
 require("../js/analyzer.js");
-
+require("../js/preprocessor.js");
+require("../js/analyzescorer.js");
+require("../js/stringhelper.js");
 var firstParagraphArg = {
     textString: "<p>this is a default text to test the first paragraph for the word keyword</p> if the keyword is mentioned after the first paragraph, it shouldn't count the keyword. ",
     queue: ["firstParagraph"],
@@ -39,7 +41,21 @@ var firstParagraphArg3 =  {
 describe("a test for checking the first paragraph for keywords, when no paragraph is defined", function(){
     it("returns 0, since there are 2 paragraphs and 2nd matches keyword", function(){
         firstParagraphAnalyzer3 = new Analyzer(firstParagraphArg3);
-        result = firstParagraphAnalyzer2.firstParagraph();
+        result = firstParagraphAnalyzer3.firstParagraph();
         expect(result[0].result).toBe(0);
+    });
+});
+
+var firstParagraphArg4 = {
+    textString: "One question we get quite often in our website reviews \n\n is whether we can help people recover from the drop they noticed in their rankings or traffic. A lot of the times, this is a legitimate drop and people were actually in a bit of trouble.",
+    keyword: "website",
+    queue: ["firstParagraph"]
+};
+
+describe("a test for checking the firstparagraph with newlines in stead of <p>-tags", function(){
+    it("returns 1, the keyword is used before the double linebreak", function(){
+        firstParagraphAnalyzer4 = new Analyzer(firstParagraphArg4);
+        result = firstParagraphAnalyzer4.firstParagraph();
+        expect(result[0].result).toBe(1);
     });
 });
