+++ spec/pagetitleKeywordSpec.js
@@ -1,6 +1,9 @@
 require("../js/config/config.js");
-require("../js/analyzer.js");
 require("../js/config/scoring.js");
+require("../js/analyzer.js");
+require("../js/preprocessor.js");
+require("../js/analyzescorer.js");
+require("../js/stringhelper.js");
 
 titleKeywordArg = {
     pageTitle: "this is a pagetitle",
